Added styles to styles.css and edited index.html
I added colors and different fonts to styles.css, which is linked to in
index.html. I also tried reordering some of the website, but I have not
yet figured out how to put the image in the center of index.html. That,
however, will be for a different day.